Fix events without a timezone displaying incorrect time.
The calendar script internally treats the local time components of a Date as being the UTC time of an event (eg, even if a Date object prints 19:00 CEST, we interpret that as 19:00 UTC).
Calendar events where the start times were specified in UTC or as 'floating' times weren't being converted into this internal way of representing the time of an event. This has been fixed.
Floating times will be fixed to the server's timezone, as we have no way of displaying such representations in Discord.